Added filters and delete confirmation to CustomerList

+++ src/components/CustomerList.jsx
@@ -13,7 +13,7 @@ import EditCustomerForm from "./EditCustomerForm";
 
 export default function CustomerList() {
 
-	
+
 	const gridRef = useRef(null);
 
 	const [customers, setCustomers] = useState([{
@@ -29,10 +29,10 @@ export default function CustomerList() {
 
 	//ag-grid taulukon sarakkeet 
 	const [colDefs, setColDefs] = useState([
-		{ field: 'firstname', flex: 1 },
-		{ field: 'lastname', flex: 1 },
-		{ field: 'email', flex: 1 },
-		{ field: 'phone', flex: 1 },
+		{ field: 'firstname', flex: 1, sortable: true, filter: true, floatingFilter: true },
+		{ field: 'lastname', flex: 1, sortable: true, filter: true, floatingFilter: true },
+		{ field: 'email', flex: 1, sortable: true, filter: true, floatingFilter: true },
+		{ field: 'phone', flex: 1, sortable: true, filter: true, floatingFilter: true },
 		{
 			headerName: '',
 			cellRenderer: (params) =>
@@ -50,7 +50,7 @@ export default function CustomerList() {
 				<EditCustomerForm func={saveEditedCustomer} params={params} />
 			, flex: 1,
 		}
-	
+
 
 	]);
 
@@ -84,8 +84,8 @@ export default function CustomerList() {
 				},
 				body: JSON.stringify(customer)
 			})
-			.then(getCustomers() )
-			.catch(err => console.error(err) )
+			.then(getCustomers())
+			.catch(err => console.error(err))
 	}
 
 	const saveEditedCustomer = (customer) => {
@@ -98,29 +98,31 @@ export default function CustomerList() {
 				},
 				body: JSON.stringify(customer)
 			})
-			.then(getCustomers() )
-			.catch(err => console.error(err) )
+			.then(getCustomers())
+			.catch(err => console.error(err))
 	}
 
 
 	const deleteCustomer = (params) => {
 		console.log("params ", params.data._links.self.href);
+		if (window.confirm("Are you sure you want to delete this customer?")) {
 
-		fetch(params.data._links.self.href,
-			{ method: 'DELETE' })
-			.then(response => {
-				if (response.ok) {
-					setOpenSnackbar(true);
-					setMsg("Delete succeed");
-					getCustomers();
-				}
-				else {
-					openSnackbar(false);
-				}
-			})
-			.catch(err => {
-				// Something went wrong
-			});
+			fetch(params.data._links.self.href,
+				{ method: 'DELETE' })
+				.then(response => {
+					if (response.ok) {
+						setOpenSnackbar(true);
+						setMsg("Delete succeed");
+						getCustomers();
+					}
+					else {
+						openSnackbar(false);
+					}
+				})
+				.catch(err => {
+					// Something went wrong
+				});
+		}
 	}
 
 
@@ -143,12 +145,12 @@ export default function CustomerList() {
 			{
 				fileName: "customers.csv",
 				columnKeys: ["firstname", "lastname", "streetaddress", "postcode",
-		"city","email", "phone"]
+					"city", "email", "phone"]
 			}
 		);
 	}
 
-	
+
 	//näytä autot nettisivulla 
 	return (
 		<>
@@ -164,8 +166,8 @@ export default function CustomerList() {
 				</AgGridReact>
 
 				<Button onClick={handleExportOnClick}>Export customers as CSV</Button>
-				<AddCustomerForm func={addCustomerFunc}/>
-				
+				<AddCustomerForm func={addCustomerFunc} />
+
 
 				<Snackbar
 					open={openSnackbar}
@@ -173,7 +175,7 @@ export default function CustomerList() {
 					autoHideDuration={3000}
 					onClose={() => setOpenSnackbar(false)}
 				/>
-				
+
 			</div>
 		</>
 	)
